debug: 18-feb-25 1112pm: creacion de VFDthhread
 sustituimos las maquinarias principales de los los de la maquinaria VFD  y creamos una unidad de compilacion especializada para el thread de VFD. y empezamos la depuracion
